changed activity class names
