+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/MainActivity.java
@@ -1,42 +1,79 @@
 package com.faanggang.wisetrack;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
-import android.widget.ListView;
 
-import java.util.ArrayList;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.auth.FirebaseUser;
+
+import com.google.firebase.auth.FirebaseAuth;
 
 public class MainActivity extends AppCompatActivity {
-
-    ListView experimentList;
-    ArrayAdapter<Experiment> experimentAdapter;
-    ArrayList<Experiment> experimentDataList;
+    private FirebaseAuth mAuth;
+    private FirebaseUser currentUser;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-//        setContentView(R.layout.activity_main);
+        setContentView(R.layout.start_screen);
+        mAuth = FirebaseAuth.getInstance();
 
+    }
 
-        setContentView(R.layout.main_menu);  // EDIT LATER: display main_menu later
-        // create class to handle menu button clicks?
+    @Override
+    public void onStart() {
+        super.onStart();
+        currentUser = mAuth.getCurrentUser();
+    }
 
+    public void menuClick(View view) {
+        if(currentUser == null) {
+            createNewUser();
+        }
+        else {
+            Log.w("USERID", currentUser.getUid());
+        }
 
-        Button publishButton = findViewById(R.id.menuPublish_button);
+        Intent intent = new Intent(this, MainMenuActivity.class);
+        startActivity(intent);
 
-        publishButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
+    }
 
-                Intent intent = new Intent(MainActivity.this, PublishExperimentActivity.class);
-                startActivity(intent);
-            }
-        });
+    public void getIDButton(View view) {
+        if(currentUser == null){
+            createNewUser();
+        } else {
+            Log.w("USERID", currentUser.getUid());
+        }
 
+    }
+
+    public void createNewUser(){
+        mAuth.signInAnonymously()
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            Log.d("SIGNIN", "success");
+                            FirebaseUser user = mAuth.getCurrentUser();
+                        } else {
+                            Log.w("SIGNIN", "failure");
+                        }
+                    }
+                });
     }
 }
