+++ WiseTrack/app/src/main/java/com/faanggang/wisetrack/MainMenuActivity.java
@@ -1,0 +1,40 @@
+package com.faanggang.wisetrack;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ArrayAdapter;
+import android.widget.Button;
+import android.widget.ListView;
+
+import java.util.ArrayList;
+
+public class MainMenuActivity extends AppCompatActivity {
+
+    ListView experimentList;
+    ArrayAdapter<Experiment> experimentAdapter;
+    ArrayList<Experiment> experimentDataList;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+
+        setContentView(R.layout.main_menu);
+
+
+        Button publishButton = findViewById(R.id.menuPublish_button);
+
+        publishButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                Intent intent = new Intent(MainMenuActivity.this, PublishExperimentActivity.class);
+                startActivity(intent);
+            }
+        });
+
+    }
+}
